python_server/requirements.txt
aiohttp==3.7.3
python_socketio==4.6.1